Project decline functions

+++ karaage/applications/views/admin.py
@@ -23,13 +23,18 @@ from django.core.urlresolvers import reverse
 from django.contrib import messages
 from django.db.models import Q
 from django.core.paginator import Paginator, InvalidPage, EmptyPage
+from django.core.mail import send_mail
+from django.conf import settings
+
 from andsome.util.filterspecs import Filter, FilterBar
+from andsome.forms import EmailForm
 
 from karaage.people.models import Person
 from karaage.applications.models import UserApplication, ProjectApplication, Applicant, Application
 from karaage.applications.forms import AdminInviteUserApplicationForm, ApplicantForm, LeaderApproveUserApplicationForm, AdminApproveProjectApplicationForm
-from karaage.applications.emails import send_user_invite_email, send_account_approved_email, send_account_declined_email, send_project_approved_email
+from karaage.applications.emails import send_user_invite_email, send_account_approved_email, send_account_declined_email, send_project_approved_email, render_email
 from karaage.util import log_object as log
+
 
 @permission_required('applications.add_userapplication')
 def send_invitation(request):
@@ -137,12 +142,20 @@ def decline_userapplication(request, application_id):
     if application.state != Application.WAITING_FOR_ADMIN:
         raise Http404
     if request.method == 'POST':
-        send_account_declined_email(application)
-        application.delete()
-        log(request.user, application, 3, 'Application declined')
-        return HttpResponseRedirect(reverse('kg_application_list'))
+        form = EmailForm(request.POST)
+        if form.is_valid():
+            to_email = application.applicant.email
+            subject, body = form.get_data()
+            log(request.user, application, 3, 'Application declined')
+            application.delete()
+            send_mail(subject, body, settings.ACCOUNTS_EMAIL, [to_email], fail_silently=False)
+            return HttpResponseRedirect(reverse('kg_application_list'))
+    else:
+        subject, body = render_email('account_declined', { 'receiver': application.applicant, 'project': application.project })
+        initial_data = {'body': body, 'subject': subject,}
+        form = EmailForm(initial=initial_data)
 
-    return render_to_response('applications/confirm_decline.html', {'application': application}, context_instance=RequestContext(request))
+    return render_to_response('applications/confirm_decline.html', {'application': application, 'form': form}, context_instance=RequestContext(request))
 
 
 @permission_required('applications.change_application')
@@ -169,17 +182,26 @@ def approve_projectapplication(request, application_id):
 
 @permission_required('applications.delete_application')
 def decline_projectapplication(request, application_id):
-    application = get_object_or_404(UserApplication, pk=application_id)
+    application = get_object_or_404(ProjectApplication, pk=application_id)
 
     if application.state != Application.WAITING_FOR_ADMIN:
         raise Http404
     if request.method == 'POST':
-        send_account_declined_email(application)
-        application.delete()
-        log(request.user, application, 3, 'Application declined')
-        return HttpResponseRedirect(reverse('kg_application_list'))
+        form = EmailForm(request.POST)
+        if form.is_valid():
+            to_email = application.applicant.email
+            subject, body = form.get_data()
+            log(request.user, application, 3, 'Application declined')
+            application.delete()
+            send_mail(subject, body, settings.ACCOUNTS_EMAIL, [to_email], fail_silently=False)
+            return HttpResponseRedirect(reverse('kg_application_list'))
 
-    return render_to_response('applications/project_confirm_decline.html', {'application': application}, context_instance=RequestContext(request))
+    else:
+        subject, body = render_email('project_declined', { 'receiver': application.applicant })
+        initial_data = {'body': body, 'subject': subject,}
+        form = EmailForm(initial=initial_data)
+
+    return render_to_response('applications/project_confirm_decline.html', {'application': application, 'form': form}, context_instance=RequestContext(request))
